convert SectionsList to ts
