--- src/client/apps/edit/components/content/section_list/index.styl
@@ -1,12 +1,0 @@
-@import '../../stylus_lib'
-
-.edit-sections__list
-  width 100%
-
-[data-layout='standard']
-  max-width standard-max-width
-  margin 0 auto
-  .edit-sections__list
-    max-width standard-overflow-w-margin
-  @media (max-width: 1380px)
-    margin 0 40px
