+++ src/client/apps/edit/index.styl
@@ -4,7 +4,6 @@
 @import './components/mobile'
 
 @import './components/content/section_controls'
-@import './components/content/section_list'
 @import './components/content/sections/header'
 @import './components/content/sections/related_articles'
 @import './components/content/sections/slideshow'
